Fix time measures with the tsc counter + misc
We use the rdtsc instruction to get timestamps when logging frequency. In the previous version, the values reported by rdtsc were considered as nanoseconds. However, this is not true. The tsc clock increases at a constant frequency but is not in nanoseconds. This frequency is available with the tsc_khz module in /sys/devices/system/cpu/tsc_khz. If the module is not available, the base frequency will be used, i.e. old behavior.

Minor changes to plotting are also in this commit. Each point is the median of a 3-point window. This avoids strange spikes due to bad values in the frequency counter.